+++ UIComponents.Models/Models/Inputs/UICInputSelectlist.cs
@@ -76,6 +76,30 @@ namespace UIComponents.Models.Models.Inputs
         #endregion
 
         #region Methods
+        public UICInputSelectList AddSource(UICHtmlStorage storage)
+        {
+            var source = new UICInputSelectListHtmlStorage(this, storage);
+            return AddSource(source);
+        }
+        public UICInputSelectList AddSource(UICHtmlStorage storage, Action<UICInputSelectListHtmlStorage> action)
+        {
+            var source = new UICInputSelectListHtmlStorage(this, storage);
+            action(source);
+            return AddSource(source);
+        }
+        public UICInputSelectList AddSource(out UICInputSelectListHtmlStorage outStorage, UICHtmlStorage storage)
+        {
+            outStorage = new UICInputSelectListHtmlStorage(this, storage);
+            return AddSource(outStorage);
+        }
+
+        public UICInputSelectList AddSource(UICInputSelectListHtmlStorage storedSource)
+        {
+            storedSource.InputSelectList = this;
+            ScriptCollection.AddToScripts(storedSource);
+            return this;
+        }
+
         public UICInputSelectList AddSource(UICActionGetPost source)
         {
             var selectSource = new UICInputSelectListSource(this, source);