+++ UIComponents.Web/Extensions/UICExtensions.cs
@@ -48,10 +48,8 @@ public static class UICExtensions
             return null;
         return await result.InvokeAsync(component);
     }
-    public static Task<IHtmlContent> InvokeAsync(this IViewComponentHelper component, params IUIComponent[] UIC)
-    {
-        return UIC.InvokeAsync(component);
-    }
+    public static Task<IHtmlContent> InvokeAsync(this IViewComponentHelper component, params IUIComponent[] UIC) => UIC.InvokeAsync(component);
+    public static Task<IHtmlContent> InvokeAsync(this IViewComponentHelper component, IEnumerable<IUIComponent> UIC) => UIC.InvokeAsync(component);
 
 
     public static string GetHtmlAttributes(this Dictionary<string, string> dictionary)